Rename global_mapper_params file to params.
